cohorts/2026/workshops/dlt/my-dlt-pipeline/pyproject.toml
[project]
name = "my-dlt-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dlt[workspace]>=1.22.0",
    "ibis-framework[duckdb]>=12.0.0",
    "marimo>=0.19.11",
    "pandas>=3.0.0",
    "pyarrow-hotfix>=0.7",
]
